Move PADMMConfig into PADMMSolver
