+++ newton/_src/solvers/kamino/solvers/padmm/solver.py
@@ -21,10 +21,15 @@ This is the highest-level interface for the Proximal-ADMM solver for constrained
 See the :mod:`newton._src.solvers.kamino.solvers.padmm` module for a detailed description and usage example.
 """
 
+from __future__ import annotations
+
 import math
+from dataclasses import dataclass
+from typing import Any, Literal
 
 import warp as wp
 
+from .....sim import Model, ModelBuilder
 from ...core.data import DataKamino
 from ...core.model import ModelKamino
 from ...core.size import SizeKamino
@@ -56,7 +61,7 @@ from .kernels import (
     make_initialize_solver_kernel,
     make_update_dual_variables_and_compute_primal_dual_residuals,
 )
-from .types import PADMMConfig, PADMMConfigStruct, PADMMData, PADMMPenaltyUpdate, PADMMWarmStartMode
+from .types import PADMMConfigStruct, PADMMData, PADMMPenaltyUpdate, PADMMWarmStartMode
 
 ###
 # Module interface
@@ -94,10 +99,304 @@ class PADMMSolver:
     - can be configured with various tolerances, penalty parameters, and other settings.
     """
 
+    @dataclass
+    class Config:
+        """
+        A container to bundle PADMM solver config.
+
+        Attributes:
+            primal_tolerance (float): The target tolerance on the total primal residual `r_primal`.\n
+                Must be greater than zero. Defaults to `1e-6`.
+            dual_tolerance (float): The target tolerance on the total dual residual `r_dual`.\n
+                Must be greater than zero. Defaults to `1e-6`.
+            compl_tolerance (float): The target tolerance on the total complementarity residual `r_compl`.\n
+                Must be greater than zero. Defaults to `1e-6`.
+            restart_tolerance (float): The tolerance on the total combined primal-dual
+                residual `r_comb`, for determining when gradient acceleration should be restarted.\n
+                Must be greater than zero. Defaults to `0.999`.
+            eta (float): The proximal regularization parameter.\n
+                Must be greater than zero. Defaults to `1e-5`.
+            rho_0 (float): The initial value of the ALM penalty parameter.\n
+                Must be greater than zero. Defaults to `1.0`.
+            rho_min (float): The lower-bound applied to the ALM penalty parameter.\n
+                Used to ensure numerical stability when adaptive penalty updates are used.\n
+                Must be greater than zero. Defaults to `1e-5`.
+            a_0 (float): The initial value of the acceleration parameter.\n
+                Must be greater than zero. Defaults to `1.0`.
+            alpha (float): The threshold on primal-dual residual ratios,
+                used to determine when penalty updates should occur.\n
+                Must be greater than `1.0`. Defaults to `10.0`.
+            tau (float): The factor by which the penalty is increased/decreased
+                when the primal-dual residual ratios exceed the threshold `alpha`.\n
+                Must be greater than `1.0`. Defaults to `1.5`.
+            max_iterations (int): The maximum number of solver iterations.\n
+                Must be greater than zero. Defaults to `200`.
+            penalty_update_freq (int): The permitted frequency of penalty updates.\n
+                If zero, no updates are performed. Otherwise, updates are performed every
+                `penalty_update_freq` iterations. Defaults to `10`.
+            penalty_update_method (str): The penalty update method used to adapt the penalty parameter.\n
+                Defaults to `fixed`. See :class:`PADMMPenaltyUpdate` for details.
+
+        """
+
+        primal_tolerance: float = 1e-6
+        """
+        The target tolerance on the total primal residual `r_primal`.\n
+        Must be greater than zero. Defaults to `1e-6`.
+        """
+
+        dual_tolerance: float = 1e-6
+        """
+        The target tolerance on the total dual residual `r_dual`.\n
+        Must be greater than zero. Defaults to `1e-6`.
+        """
+
+        compl_tolerance: float = 1e-6
+        """
+        The target tolerance on the total complementarity residual `r_compl`.\n
+        Must be greater than zero. Defaults to `1e-6`.
+        """
+
+        restart_tolerance: float = 0.999
+        """
+        The tolerance on the total combined primal-dual residual `r_comb`,
+        for determining when gradient acceleration should be restarted.\n
+        Must be greater than zero. Defaults to `0.999`.
+        """
+
+        eta: float = 1e-5
+        """
+        The proximal regularization parameter.\n
+        Must be greater than zero. Defaults to `1e-5`.
+        """
+
+        rho_0: float = 1.0
+        """
+        The initial value of the ALM penalty parameter.\n
+        Must be greater than zero. Defaults to `1.0`.
+        """
+
+        rho_min: float = 1e-5
+        """
+        The lower-bound applied to the ALM penalty parameter.\n
+        Used to ensure numerical stability when adaptive penalty updates are used.\n
+        Must be greater than zero. Defaults to `1e-5`.
+        """
+
+        a_0: float = 1.0
+        """
+        The initial value of the acceleration parameter.\n
+        Must be greater than zero. Defaults to `1.0`.
+        """
+
+        alpha: float = 10.0
+        """
+        The primal-dual residual threshold used to determine when penalty updates are needed.
+        Must be greater than one. Defaults to `10.0`.
+        """
+
+        tau: float = 1.5
+        """
+        The factor by which the ALM penalty is increased/decreased when
+        the primal-dual residual ratios exceed the threshold `alpha`.\n
+        Must be greater than `1.0`. Defaults to `1.5`.
+        """
+
+        max_iterations: int = 200
+        """
+        The maximum number of solver iterations.\n
+        Must be greater than zero. Defaults to `200`.
+        """
+
+        penalty_update_freq: int = 1
+        """
+        The permitted frequency of penalty updates.\n
+        If zero, no updates are performed. Otherwise, updates are performed every
+        `penalty_update_freq` iterations. Defaults to `1`.
+        """
+
+        penalty_update_method: Literal["fixed", "balanced"] = "fixed"
+        """
+        The penalty update method used to adapt the penalty parameter.\n
+        Defaults to `fixed`. See :class:`PADMMPenaltyUpdate` for details.
+        """
+
+        linear_solver_tolerance: float = 0.0
+        """
+        The default absolute tolerance for the iterative linear solver.\n
+        When positive, the iterative solver's atol is initialized to this value
+        at the start of each ADMM solve.\n
+        When zero, the iterative solver's own tolerance is left unchanged.\n
+        Must be non-negative. Defaults to `0.0`.
+        """
+
+        linear_solver_tolerance_ratio: float = 0.0
+        """
+        The ratio used to adapt the iterative linear solver tolerance from the ADMM primal residual.\n
+        When positive, the linear solver absolute tolerance is set to
+        `ratio * ||r_primal||_2` at each ADMM iteration.\n
+        When zero, the linear solver tolerance is not adapted (fixed tolerance).\n
+        Must be non-negative. Defaults to `0.0`.
+        """
+
+        def check(self):
+            """
+            Checks the validity of PADMM solver config.
+            """
+            if self.primal_tolerance < 0.0:
+                raise ValueError(f"Invalid primal tolerance: {self.primal_tolerance}. Must be non-negative.")
+            if self.dual_tolerance < 0.0:
+                raise ValueError(f"Invalid dual tolerance: {self.dual_tolerance}. Must be non-negative.")
+            if self.compl_tolerance < 0.0:
+                raise ValueError(f"Invalid complementarity tolerance: {self.compl_tolerance}. Must be non-negative.")
+            if not (0.0 <= self.restart_tolerance < 1.0):
+                raise ValueError(
+                    f"Invalid restart tolerance: {self.restart_tolerance}. Must be in the range [0.0, 1.0)."
+                )
+            if self.eta <= 0.0:
+                raise ValueError(f"Invalid proximal parameter: {self.eta}. Must be greater than zero.")
+            if self.rho_0 <= 0.0:
+                raise ValueError(f"Invalid initial ALM penalty: {self.rho_0}. Must be greater than zero.")
+            if self.rho_min <= 0.0:
+                raise ValueError(f"Invalid minimum ALM penalty: {self.rho_min}. Must be greater than zero.")
+            if self.a_0 <= 0.0:
+                raise ValueError(f"Invalid initial acceleration parameter: {self.a_0}. Must be greater than zero.")
+            if self.alpha <= 1.0:
+                raise ValueError(f"Invalid penalty threshold: {self.alpha}. Must be greater than one.")
+            if self.tau <= 1.0:
+                raise ValueError(f"Invalid penalty increment factor: {self.tau}. Must be greater than one.")
+            if self.max_iterations <= 0:
+                raise ValueError(f"Invalid maximum iterations: {self.max_iterations}. Must be a positive integer.")
+            if self.penalty_update_freq < 0:
+                raise ValueError(f"Invalid penalty update frequency: {self.penalty_update_freq}. Must be non-negative.")
+            # Conversion to PADMMPenaltyUpdate will raise an error if the input string is invalid.
+            PADMMPenaltyUpdate.from_string(self.penalty_update_method)
+            if self.linear_solver_tolerance < 0.0:
+                raise ValueError(
+                    f"Invalid linear solver tolerance: {self.linear_solver_tolerance}. Must be non-negative."
+                )
+            if self.linear_solver_tolerance_ratio < 0.0:
+                raise ValueError(
+                    f"Invalid linear solver tolerance ratio: {self.linear_solver_tolerance_ratio}. Must be non-negative."
+                )
+
+        def to_struct(self) -> PADMMConfigStruct:
+            """
+            Converts the host-side config to the corresponding device-side object.
+
+            Returns:
+                PADMMConfigStruct: The solver config as a warp struct.
+            """
+            config_struct = PADMMConfigStruct()
+            config_struct.primal_tolerance = self.primal_tolerance
+            config_struct.dual_tolerance = self.dual_tolerance
+            config_struct.compl_tolerance = self.compl_tolerance
+            config_struct.restart_tolerance = self.restart_tolerance
+            config_struct.eta = self.eta
+            config_struct.rho_0 = self.rho_0
+            config_struct.rho_min = self.rho_min
+            config_struct.a_0 = self.a_0
+            config_struct.alpha = self.alpha
+            config_struct.tau = self.tau
+            config_struct.max_iterations = self.max_iterations
+            config_struct.penalty_update_freq = self.penalty_update_freq
+            config_struct.penalty_update_method = PADMMPenaltyUpdate.from_string(self.penalty_update_method)
+            config_struct.linear_solver_tolerance = self.linear_solver_tolerance
+            config_struct.linear_solver_tolerance_ratio = self.linear_solver_tolerance_ratio
+            return config_struct
+
+        def __post_init__(self):
+            """Post-initialization to validate config."""
+            self.check()
+
+        @classmethod
+        def register_custom_attributes(cls, builder: ModelBuilder) -> None:
+            """
+            Register custom attributes for this config.
+
+            Args:
+                builder (ModelBuilder): The model builder to register the custom attributes to.
+            """
+
+            # Register KaminoSceneAPI attributes so the USD importer will store them on the model
+            builder.add_custom_attribute(
+                ModelBuilder.CustomAttribute(
+                    name="padmm_primal_tolerance",
+                    frequency=Model.AttributeFrequency.ONCE,
+                    assignment=Model.AttributeAssignment.MODEL,
+                    dtype=wp.float32,
+                    default=1e-6,
+                    namespace="kamino",
+                    usd_attribute_name="newton:kamino:padmm:primalTolerance",
+                )
+            )
+            builder.add_custom_attribute(
+                ModelBuilder.CustomAttribute(
+                    name="padmm_dual_tolerance",
+                    frequency=Model.AttributeFrequency.ONCE,
+                    assignment=Model.AttributeAssignment.MODEL,
+                    dtype=wp.float32,
+                    default=1e-6,
+                    namespace="kamino",
+                    usd_attribute_name="newton:kamino:padmm:dualTolerance",
+                )
+            )
+            builder.add_custom_attribute(
+                ModelBuilder.CustomAttribute(
+                    name="padmm_complementarity_tolerance",
+                    frequency=Model.AttributeFrequency.ONCE,
+                    assignment=Model.AttributeAssignment.MODEL,
+                    dtype=wp.float32,
+                    default=1e-6,
+                    namespace="kamino",
+                    usd_attribute_name="newton:kamino:padmm:complementarityTolerance",
+                )
+            )
+
+            # Separately register `newton:maxSolverIterations` from `KaminoSceneAPI` so we have access
+            # to it through the model.
+            builder.add_custom_attribute(
+                ModelBuilder.CustomAttribute(
+                    name="max_solver_iterations",
+                    frequency=Model.AttributeFrequency.ONCE,
+                    assignment=Model.AttributeAssignment.MODEL,
+                    dtype=wp.int32,
+                    default=-1,
+                    namespace="kamino",
+                    usd_attribute_name="newton:maxSolverIterations",
+                )
+            )
+
+        @staticmethod
+        def from_model(model: Model, **kwargs: dict[str, Any]) -> PADMMSolver.Config:
+            """Creates a config based on a model, using any config parameters that might be stored in
+            the model if it was imported from USD.
+
+            Args:
+                model: Newton model.
+            """
+            config = PADMMSolver.Config(**kwargs)
+
+            # Parse solver-specific attributes imported from USD
+            kamino_attrs = getattr(model, "kamino", None)
+            if kamino_attrs is not None:
+                if hasattr(kamino_attrs, "padmm_primal_tolerance"):
+                    config.primal_tolerance = float(kamino_attrs.padmm_primal_tolerance.numpy()[0])
+                if hasattr(kamino_attrs, "padmm_dual_tolerance"):
+                    config.dual_tolerance = float(kamino_attrs.padmm_dual_tolerance.numpy()[0])
+                if hasattr(kamino_attrs, "padmm_complementarity_tolerance"):
+                    config.compl_tolerance = float(kamino_attrs.padmm_complementarity_tolerance.numpy()[0])
+                if hasattr(kamino_attrs, "max_solver_iterations"):
+                    max_iterations = kamino_attrs.max_solver_iterations.numpy()[0]
+                    if max_iterations >= 0:
+                        config.max_iterations = max_iterations
+
+            return config
+
     def __init__(
         self,
         model: ModelKamino | None = None,
-        config: list[PADMMConfig] | PADMMConfig | None = None,
+        config: list[PADMMSolver.Config] | PADMMSolver.Config | None = None,
         warmstart: PADMMWarmStartMode = PADMMWarmStartMode.NONE,
         use_acceleration: bool = True,
         collect_info: bool = False,
@@ -114,7 +413,7 @@ class PADMMSolver:
             model (ModelKamino | None): The model for which to allocate the solver data.
             limits (LimitsKamino | None): The limits container associated with the model.
             contacts (ContactsKamino | None): The contacts container associated with the model.
-            config (list[PADMMConfig] | PADMMConfig | None): The solver config to use.
+            config (list[PADMMSolver.Config] | PADMMSolver.Config | None): The solver config to use.
             use_acceleration (bool): Set to `True` to enable Nesterov acceleration.
             collect_info (bool): Set to `True` to enable collection of solver convergence info.\n
                 This setting is intended only for analysis and debugging purposes, as it
@@ -126,7 +425,7 @@ class PADMMSolver:
 
         # Declare the internal solver config cache
         self._warmstart: PADMMWarmStartMode = PADMMWarmStartMode.NONE
-        self._config: list[PADMMConfig] = []
+        self._config: list[PADMMSolver.Config] = []
         self._use_acceleration: bool = False
         self._collect_info: bool = False
         self._avoid_graph_conditionals: bool = False
@@ -158,10 +457,10 @@ class PADMMSolver:
     ###
 
     @property
-    def config(self) -> list[PADMMConfig]:
+    def config(self) -> list[PADMMSolver.Config]:
         """
         Returns the host-side cache of the solver config.\n
-        They are used to construct the warp array of type :class:`PADMMConfigStruct` on the target device.
+        They are used to construct the warp array of type :class:`PADMMSolver.Config` on the target device.
         """
         return self._config
 
@@ -195,7 +494,7 @@ class PADMMSolver:
     def finalize(
         self,
         model: ModelKamino | None = None,
-        config: list[PADMMConfig] | PADMMConfig | None = None,
+        config: list[PADMMSolver.Config] | PADMMSolver.Config | None = None,
         warmstart: PADMMWarmStartMode = PADMMWarmStartMode.NONE,
         use_acceleration: bool = True,
         collect_info: bool = False,
@@ -209,7 +508,7 @@ class PADMMSolver:
             model (ModelKamino | None): The model for which to allocate the solver data.
             limits (LimitsKamino | None): The limits container associated with the model.
             contacts (ContactsKamino | None): The contacts container associated with the model.
-            config (list[PADMMConfig] | PADMMConfig | None): The solver config to use.
+            config (list[PADMMSolver.Config] | PADMMSolver.Config | None): The solver config to use.
             use_acceleration (bool): Set to `True` to enable Nesterov acceleration.
             collect_info (bool): Set to `True` to enable collection of solver convergence info.\n
                 This setting is intended only for analysis and debugging purposes, as it
@@ -393,39 +692,39 @@ class PADMMSolver:
 
     @staticmethod
     def _check_config(
-        model: ModelKamino | None = None, config: list[PADMMConfig] | PADMMConfig | None = None
-    ) -> list[PADMMConfig]:
+        model: ModelKamino | None = None, config: list[PADMMSolver.Config] | PADMMSolver.Config | None = None
+    ) -> list[PADMMSolver.Config]:
         """
         Checks and validates the provided solver config, returning a list
         of config objects corresponding to each world in the model.
 
         Args:
             model (ModelKamino | None): The model for which to validate the config.
-            config (list[PADMMConfig] | PADMMConfig | None): The solver config to validate.
+            config (list[PADMMSolver.Config] | PADMMSolver.Config | None): The solver config to validate.
         """
         # If no config is provided, use defaults
         if config is None:
             # If no model is provided, use a single default config object
             if model is None:
-                config = [PADMMConfig()]
+                config = [PADMMSolver.Config()]
 
             # If a model is provided, create a list of default config
             # objects based on the number of worlds in the model
             else:
                 num_worlds = model.info.num_worlds
-                config = [PADMMConfig()] * num_worlds
+                config = [PADMMSolver.Config()] * num_worlds
 
         # If a single config object is provided, convert it to a list
-        elif isinstance(config, PADMMConfig):
+        elif isinstance(config, PADMMSolver.Config):
             config = [config] * (model.info.num_worlds if model else 1)
 
         # If a list of configs is provided, ensure it matches the number
-        # of worlds and that all configs are instances of PADMMConfig
+        # of worlds and that all configs are instances of PADMMSolver.Config
         elif isinstance(config, list):
             if model is not None and len(config) != model.info.num_worlds:
                 raise ValueError(f"Expected {model.info.num_worlds} configs, got {len(config)}")
-            if not all(isinstance(s, PADMMConfig) for s in config):
-                raise TypeError("All configs must be instances of PADMMConfig")
+            if not all(isinstance(s, PADMMSolver.Config) for s in config):
+                raise TypeError("All configs must be instances of PADMMSolver.Config")
 
         # Return the validated config
         return config
