+++ newton/_src/solvers/kamino/tests/test_solvers_padmm.py
@@ -30,7 +30,7 @@ from newton._src.solvers.kamino.linalg import ConjugateResidualSolver, LLTBlocke
 from newton._src.solvers.kamino.linalg.utils.matrix import SquareSymmetricMatrixProperties
 from newton._src.solvers.kamino.linalg.utils.range import in_range_via_gaussian_elimination
 from newton._src.solvers.kamino.models.builders import basics
-from newton._src.solvers.kamino.solvers.padmm import PADMMConfig, PADMMSolver, PADMMWarmStartMode
+from newton._src.solvers.kamino.solvers.padmm import PADMMSolver, PADMMWarmStartMode
 from newton._src.solvers.kamino.tests import setup_tests, test_context
 from newton._src.solvers.kamino.tests.utils.extract import (
     extract_delassus,
@@ -400,7 +400,7 @@ class TestPADMMSolver(unittest.TestCase):
         # NOTE: These are all equal to their default values
         # but are defined here explicitly for the purposes
         # of experimentation and testing.
-        config = PADMMConfig()
+        config = PADMMSolver.Config()
         config.primal_tolerance = 1e-6
         config.dual_tolerance = 1e-6
         config.compl_tolerance = 1e-6
@@ -444,7 +444,7 @@ class TestPADMMSolver(unittest.TestCase):
         # NOTE: These are all equal to their default values
         # but are defined here explicitly for the purposes
         # of experimentation and testing.
-        config = PADMMConfig()
+        config = PADMMSolver.Config()
         config.primal_tolerance = 1e-6
         config.dual_tolerance = 1e-6
         config.compl_tolerance = 1e-6
@@ -489,7 +489,7 @@ class TestPADMMSolver(unittest.TestCase):
         # NOTE: These are all equal to their default values
         # but are defined here explicitly for the purposes
         # of experimentation and testing.
-        config = PADMMConfig()
+        config = PADMMSolver.Config()
         config.primal_tolerance = 1e-6
         config.dual_tolerance = 1e-6
         config.compl_tolerance = 1e-6
@@ -536,7 +536,7 @@ class TestPADMMSolver(unittest.TestCase):
         # NOTE: These are all equal to their default values
         # but are defined here explicitly for the purposes
         # of experimentation and testing.
-        config = PADMMConfig()
+        config = PADMMSolver.Config()
         config.primal_tolerance = 1e-6
         config.dual_tolerance = 1e-6
         config.compl_tolerance = 1e-6
@@ -584,7 +584,7 @@ class TestPADMMSolver(unittest.TestCase):
         # NOTE: These are all equal to their default values
         # but are defined here explicitly for the purposes
         # of experimentation and testing.
-        config = PADMMConfig()
+        config = PADMMSolver.Config()
         config.primal_tolerance = 1e-6
         config.dual_tolerance = 1e-6
         config.compl_tolerance = 1e-6
@@ -632,7 +632,7 @@ class TestPADMMSolver(unittest.TestCase):
         # NOTE: These are all equal to their default values
         # but are defined here explicitly for the purposes
         # of experimentation and testing.
-        config = PADMMConfig()
+        config = PADMMSolver.Config()
         config.primal_tolerance = 1e-6
         config.dual_tolerance = 1e-6
         config.compl_tolerance = 1e-6
